requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
langchain==0.3.7
numpy==2.1.2
openai==1.54.4
sentence-transformers==3.2.0
fastapi==0.115.4
uvicorn==0.32.0
